requirements.txt
quantstats
pandas
selenium
webdriver_manager
MetaTrader5
streamlit
python-dotenv
numpy==1.26.4
yfinance